refactor: refactored controller and enities
