+++ src/main/java/osz/imt/mts/mts11umfrage/dto/QuestionDto.java
@@ -7,8 +7,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import osz.imt.mts.mts11umfrage.data.QuestionTypes;
 import osz.imt.mts.mts11umfrage.models.Question;
-import osz.imt.mts.mts11umfrage.utils.QuestionTypes;
 
 /**
  * A DTO for the {@link Question} entity.
