+++ src/main/java/osz/imt/mts/mts11umfrage/service/QuestionService.java
@@ -26,7 +26,7 @@ public class QuestionService {
    */
   private final QuestionAnswerRepository answerRepo;
   /**
-   * User answers repository.
+   * UserData answers repository.
    */
   private final UserAnswersRepository userAnswerRepo;
